Apply stashed changes: Use 'owner' middleware on routes to prevent unauthorized access

+++ routes/web.php
@@ -30,17 +30,17 @@ Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name
 //needs to be logged in
 Route::middleware(['auth'])->group(function () {
     //boarding house
-    Route::get('/boarding-houses/create', [BoardingHouseController::class, 'create'])->name('boarding-houses.create'); // Show form to create
-    Route::post('/boarding-houses', [BoardingHouseController::class, 'store'])->name('boarding-houses.store'); // Save new record
+    Route::get('/boarding-houses/create', [BoardingHouseController::class, 'create'])->name('boarding-houses.create')->middleware('owner'); // Show form to create
+    Route::post('/boarding-houses', [BoardingHouseController::class, 'store'])->name('boarding-houses.store')->middleware('owner'); // Save new record
     Route::get('/boarding-houses/{boardingHouse}', [BoardingHouseController::class, 'show'])->name('boarding-houses.show'); // Show details
-    Route::get('/boarding-houses/{boardingHouse}/edit', [BoardingHouseController::class, 'edit'])->name('boarding-houses.edit'); // Show form to edit
-    Route::post('/boarding-houses/{boardingHouse}', [BoardingHouseController::class, 'update'])->name('boarding-houses.update'); // Update record
-    Route::delete('/boarding-houses/{boardingHouse}', [BoardingHouseController::class, 'destroy'])->name('boarding-houses.destroy'); // Delete record
+    Route::get('/boarding-houses/{boardingHouse}/edit', [BoardingHouseController::class, 'edit'])->name('boarding-houses.edit')->middleware('owner'); // Show form to edit
+    Route::post('/boarding-houses/{boardingHouse}', [BoardingHouseController::class, 'update'])->name('boarding-houses.update')->middleware('owner'); // Update record
+    Route::delete('/boarding-houses/{boardingHouse}', [BoardingHouseController::class, 'destroy'])->name('boarding-houses.destroy')->middleware('owner'); // Delete record
 
     //rooms
     Route::get('/rooms', [RoomController::class, 'index'])->name('rooms.index'); // List all
-    Route::get('/boarding-houses/{boardingHouse}/rooms/create', [RoomController::class, 'create'])->name('rooms.create'); // Show form to create
-    Route::post('/rooms/store', [RoomController::class, 'store'])->name('rooms.store'); // Save new record
+    Route::get('/boarding-houses/{boardingHouse}/rooms/create', [RoomController::class, 'create'])->name('rooms.create')->middleware('owner'); // Show form to create
+    Route::post('/rooms/store', [RoomController::class, 'store'])->name('rooms.store')->middleware('owner'); // Save new record
     Route::get('/rooms/{room}', [RoomController::class, 'show'])->name('rooms.show'); // Show details
 
     //boarders
@@ -50,6 +50,6 @@ Route::middleware(['auth'])->group(function () {
     Route::get('/boarders', [BoarderController::class, 'show'])->name('boarders.show'); // Show details
 
     //accept or decline boarders
-    Route::post('/boarders/decision', [BoarderController::class, 'acceptOrDecline'])->name('boarders.decision'); // Show details
+    Route::post('/boarders/decision', [BoarderController::class, 'acceptOrDecline'])->name('boarders.decision')->middleware('owner'); // Show details
 
 });
